+++ app/controllers/draws_controller.rb
@@ -67,29 +67,6 @@ class DrawsController < ApplicationController # rubocop:disable ClassLength
     handle_action(**result)
   end
 
-  def suite_summary
-    suites = @draw.suites.includes(:rooms).available.where(medical: false)
-    @all_sizes = SuiteSizesQuery.new(suites).call
-    @suites_by_size = SuitesBySizeQuery.new(suites).call
-    @suites_by_size.default = []
-  end
-
-  def suites_edit
-    prepare_suites_edit_data
-  end
-
-  def suites_update
-    result = DrawSuitesUpdate.update(draw: @draw, params: suites_update_params)
-    @suites_update = result[:update_object]
-    if @suites_update
-      prepare_suites_edit_data
-      result[:action] = 'suites_edit'
-    else
-      result[:path] = suite_summary_draw_path(@draw)
-    end
-    handle_action(**result)
-  end
-
   def student_summary
     prepare_students_edit_data
   end
@@ -165,10 +142,6 @@ class DrawsController < ApplicationController # rubocop:disable ClassLength
                                  locked_sizes: [])
   end
 
-  def suites_update_params
-    params.require(:draw_suites_update).permit(suite_edit_param_hash)
-  end
-
   def students_update_params
     params.fetch(:draw_students_update, {}).permit(:class_year)
   end
@@ -187,25 +160,6 @@ class DrawsController < ApplicationController # rubocop:disable ClassLength
 
   def calculate_metrics
     @draw = DrawReport.new(@draw)
-  end
-
-  def prepare_suites_edit_data # rubocop:disable AbcSize, MethodLength
-    draw_suites = @draw.suites.available.where(medical: false)
-    all_suites = Suite.available.where(medical: false)
-    @suite_sizes ||= SuiteSizesQuery.new(all_suites).call
-    @suites_update ||= DrawSuitesUpdate.new(draw: @draw)
-    base_suites = all_suites.order(:number)
-    empty_suite_hash = @suite_sizes.map { |s| [s, []] }.to_h
-    @current_suites = empty_suite_hash.merge(
-      draw_suites.includes(:draws).order(:number).group_by(&:size)
-    )
-    @drawless_suites = empty_suite_hash.merge(
-      DrawlessSuitesQuery.new(base_suites).call.group_by(&:size)
-    )
-    @drawn_suites = empty_suite_hash.merge(
-      SuitesInOtherDrawsQuery.new(base_suites).call(draw: @draw)
-                             .group_by(&:size)
-    )
   end
 
   def prepare_students_edit_data
@@ -241,15 +195,5 @@ class DrawsController < ApplicationController # rubocop:disable ClassLength
       result[:path] = student_summary_draw_path(@draw)
     end
     result
-  end
-
-  def suite_edit_param_hash
-    suite_edit_sizes.flat_map do |s|
-      DrawSuitesUpdate::CONSOLIDATED_ATTRS.map { |p| ["#{p}_#{s}".to_sym, []] }
-    end.to_h
-  end
-
-  def suite_edit_sizes
-    @suite_edit_sizes ||= SuiteSizesQuery.new(Suite.available).call
   end
 end