+++ app/models/draw_suite.rb
@@ -2,7 +2,7 @@
 
 # Intermediary model for draws <--> suites, originally a has_and_belongs_to_many
 # association but now has_many :through
-class DrawsSuite < ApplicationRecord
+class DrawSuite < ApplicationRecord
   belongs_to :draw
   belongs_to :suite
 end