+++ spec/models/draw_spec.rb
@@ -7,8 +7,8 @@ RSpec.describe Draw, type: :model do
     it { is_expected.to validate_presence_of(:name) }
     it { is_expected.to have_many(:students) }
     it { is_expected.to have_many(:groups) }
-    it { is_expected.to have_many(:draws_suites) }
-    it { is_expected.to have_many(:suites).through(:draws_suites) }
+    it { is_expected.to have_many(:draw_suites) }
+    it { is_expected.to have_many(:suites).through(:draw_suites) }
     it { is_expected.to validate_presence_of(:status) }
     it { is_expected.to validate_presence_of(:suite_selection_mode) }