+++ spec/models/suite_spec.rb
@@ -8,8 +8,8 @@ RSpec.describe Suite, type: :model do
     it { is_expected.to belong_to(:building) }
     it { is_expected.to belong_to(:group) }
     it { is_expected.to have_many(:rooms) }
-    it { is_expected.to have_many(:draws_suites) }
-    it { is_expected.to have_many(:draws).through(:draws_suites) }
+    it { is_expected.to have_many(:draw_suites) }
+    it { is_expected.to have_many(:draws).through(:draw_suites) }
 
     describe 'number uniqueness' do
       it 'allows duplicates that belong to separate buildings' do